Fixed some of the pylint errors from codacy. Moved the get_future_datetime method to the utils file
