+++ requirements-dev.txt
@@ -1,3 +1,4 @@
+beautifulsoup4
 coverage
 freezegun
 ipdb
